added handlers

+++ internal/service/auth.go
@@ -13,7 +13,7 @@ func NewAuthService(storage storage.Auth) *AuthService {
 	return &AuthService{storage: storage}
 }
 
-func (s *AuthService) Register(ctx context.Context, email, password string) (userId int, err error) {
+func (s *AuthService) Register(ctx context.Context, email, password string) (userId int64, err error) {
 	panic("implement me")
 }
 func (s *AuthService) Login(ctx context.Context, email, password string) (token string, err error) {
