+++ internal/service/investment.go
@@ -13,7 +13,7 @@ func NewInvestmentService(storage storage.Investment) *InvestmentService {
 	return &InvestmentService{storage: storage}
 }
 
-func (s *InvestmentService) Create(ctx context.Context, amount int64, currency string) (investmentId int, err error) {
+func (s *InvestmentService) Create(ctx context.Context, amount int64, currency string) (investmentId int64, err error) {
 	panic("implement me")
 }
 
@@ -21,6 +21,6 @@ func (s *InvestmentService) Get(ctx context.Context) (amount int64, currency str
 	panic("implement me")
 }
 
-func (s *InvestmentService) Delete(ctx context.Context, investmentId int) error {
+func (s *InvestmentService) Delete(ctx context.Context, investmentId int64) error {
 	panic("implement me")
 }
