+++ internal/storage/storage.go
@@ -19,12 +19,12 @@ func NewStorage(db *pgxpool.Pool) *Storage {
 }
 
 type Auth interface {
-	Register(ctx context.Context, email, password string) (userId int, err error)
+	Register(ctx context.Context, email, password string) (userId int64, err error)
 	Login(ctx context.Context, email, password string) (token string, err error)
 }
 
 type Investment interface {
-	Create(ctx context.Context, amount int64, currency string) (investmentId int, err error)
+	Create(ctx context.Context, amount int64, currency string) (investmentId int64, err error)
 	Get(ctx context.Context) (amount int64, currency string, err error)
-	Delete(ctx context.Context, investmentId int) error
+	Delete(ctx context.Context, investmentId int64) error
 }
